integration test for a valid dataset (#252)
* integration test for valid dataset
* use np.testing.assert_allclose
* correct over-assertion
* mark as datarepo
* force conversion to np.array
* explicit extension for glob pattern of TIFF files
* Follow save_dir signature change
* Use different function to load tiffs
* fix test_run fetching of output
* change omegas to rot_angles
* Fix finding tiffs
* Add missing arguments
* add tmpdir fixture
* endow input data dirs with absolute path
* add gold data
* refactor saving the log and config
* use same conda incubator in unit and integration testing
* force download of miniconda
* avoid having two tests writing to the same output directory
* detach the filehandler before moving the log file
* sort the saved tiffiles upon loading
* migrate to fixture tmpdir
* unskip two tests

+++ src/imars3d/backend/dataio/data.py
@@ -240,7 +240,7 @@ def _load_images(filelist: List[str], desc: str, max_workers: int, tqdm_class) -
 
     rst = process_map(partial(_forgiving_reader, reader=reader), filelist, **kwargs)
     # return the results
-    return np.array([me for me in rst if me is not None])
+    return np.array([me for me in rst if me is not None], dtype="float")
 
 
 # use _func to avoid sphinx pulling it into docs
@@ -475,7 +475,8 @@ def _extract_rotation_angles(
         # img.pages[0].tags[65039].value
         # >> 'RotationActual:0.579840'
         rotation_angles = np.array(
-            [float(tifffile.TiffFile(f).pages[0].tags[metadata_idx].value.split(":")[-1]) for f in filelist]
+            [float(tifffile.TiffFile(f).pages[0].tags[metadata_idx].value.split(":")[-1]) for f in filelist],
+            dtype="float",
         )
     return rotation_angles
